requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
db-dtypes==1.0.4
decorator==5.1.1
frozenlist==1.3.1
fsspec==2022.8.2
gcsfs==2022.8.2
google-api-core==2.10.2
google-auth==2.12.0
google-auth-oauthlib==0.5.3
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
idna==3.4
multidict==6.0.2
numpy==1.21.6
oauthlib==3.2.1
packaging==21.3
pandas==1.3.5
pandas-gbq==0.17.9
proto-plus==1.22.1
protobuf==4.21.7
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
